backend/requirements.txt
# fastapi==0.111.1   # or latest compatible version
# google-auth==2.35.0
# google-auth-oauthlib==1.2.1
# google-auth-httplib2==0.2.0
# google-cloud-vision==3.8.1
# moviepy==1.0.3
# imageio==2.34.2
# decorator==4.4.2
# langchain-google-vertexai==2.0.28
# langchain-elasticsearch==0.3.2
# uvicorn==0.23.2
# google-cloud-translate==3.21.1
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
google-cloud-aiplatform==1.110.0
google-cloud-storage==2.19.0
google-cloud-vision==3.10.2
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.179.0
langchain-core==0.3.75
langchain-google-vertexai==2.0.28
rembg==2.0.67
onnxruntime==1.20.1
python-multipart==0.0.12
torch==2.8.0
torchvision==0.23.0
catboost==1.2.8
numpy==2.2.6
pillow==11.3.0
color-extraction==0.1a4
matplotlib==3.10.0
python-dotenv==1.1.1
moviepy==1.0.3
spacy==3.8.7
keybert==0.9.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
scikit-learn==1.7.1
sentence-transformers==5.1.0
transformers==4.56.0
tqdm==4.67.1
pandas==2.3.2
imageio==2.37.0
imageio-ffmpeg==0.4.9
attrs==25.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
google-adk==1.12.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.110.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.36.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.57.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.27.0
google-cloud-trace==1.16.2
google-cloud-translate==3.21.1
google-cloud-vision==3.10.2
google-crc32c==1.7.1
google-genai==1.31.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0

